loader: fix that the same app can be discovered multiple times
If an application directory is matched by multiple entries in
application_directories from the .baur.toml file, "baur ls apps" and "baur
status" would show it multiple times.

Fix it by deduplicating the list of discovered application configs.
